chore: move signature to components folder
